+++ myems-admin/app/controllers/settings/tenant/tenant.master.controller.js
@@ -1,8 +1,11 @@
 'use strict';
 
 app.controller('TenantMasterController', function($scope) {
-	$scope.$on('handleEmitTenantChanged', function(event) {
-		$scope.$broadcast('handleBroadcastTenantChanged');
-	});
-
+    $scope.$on('handleEmitTenantChanged', function(event) {
+        $scope.$broadcast('handleBroadcastTenantChanged');
+    });
+    
+    $scope.$on('handleEmitTenantTypeChanged', function(event) {
+        $scope.$broadcast('handleBroadcastTenantTypeChanged');
+    });
 });