+++ myems-api/core/datasource.py
@@ -21,6 +21,11 @@ class DataSourceCollection:
     @staticmethod
     def on_get(req, resp):
         admin_control(req)
+        search_query = req.get_param('q', default=None)
+        if search_query is not None:
+            search_query = search_query.strip()
+        else:
+            search_query = ''
         cnx = mysql.connector.connect(**config.myems_system_db)
         cursor = cnx.cursor()
 
@@ -36,9 +41,13 @@ class DataSourceCollection:
                                         "uuid": row[2]}
 
         query = (" SELECT id, name, uuid, gateway_id, protocol, connection, last_seen_datetime_utc, description "
-                 " FROM tbl_data_sources "
-                 " ORDER BY id ")
-        cursor.execute(query)
+                 " FROM tbl_data_sources " )
+        params = []
+        if search_query:
+            query += " WHERE name LIKE %s OR description LIKE %s "
+            params = [f'%{search_query}%', f'%{search_query}%']
+        query += " ORDER BY id "
+        cursor.execute(query,params)
         rows = cursor.fetchall()
         cursor.close()
         cnx.close()