+++ myems-api/core/meter.py
@@ -25,6 +25,13 @@ class MeterCollection:
             access_control(req)
         else:
             api_key_control(req)
+
+        search_query = req.get_param('q', default=None)
+        if search_query is not None:
+            search_query = search_query.strip()
+        else:
+            search_query = ''
+
         cnx = mysql.connector.connect(**config.myems_system_db)
         cursor = cnx.cursor()
 
@@ -76,12 +83,17 @@ class MeterCollection:
                                              "name": row[1],
                                              "uuid": row[2]}
 
-        query = (" SELECT id, name, uuid, energy_category_id, "
-                 "        is_counted, hourly_low_limit, hourly_high_limit, "
-                 "        cost_center_id, energy_item_id, master_meter_id, description "
-                 " FROM tbl_meters "
-                 " ORDER BY id ")
-        cursor.execute(query)
+        query_base = (" SELECT id, name, uuid, energy_category_id, "
+                      "        is_counted, hourly_low_limit, hourly_high_limit, "
+                      "        cost_center_id, energy_item_id, master_meter_id, description "
+                      " FROM tbl_meters ")
+        params = []
+        if search_query:
+            query_base += " WHERE name LIKE %s OR description LIKE %s "
+            params = [f'%{search_query}%', f'%{search_query}%']
+
+        query_base += " ORDER BY id "
+        cursor.execute(query_base, params)
         rows_meters = cursor.fetchall()
 
         result = list()