+++ myems-api/core/virtualmeter.py
@@ -25,6 +25,12 @@ class VirtualMeterCollection:
             access_control(req)
         else:
             api_key_control(req)
+        search_query = req.get_param('q', default=None)
+        if search_query is not None:
+            search_query = search_query.strip()
+        else:
+            search_query = ''
+
         cnx = mysql.connector.connect(**config.myems_system_db)
         cursor = cnx.cursor()
 
@@ -66,9 +72,14 @@ class VirtualMeterCollection:
 
         query = (" SELECT id, name, uuid, equation, energy_category_id, is_counted, cost_center_id, "
                  "        energy_item_id, description "
-                 " FROM tbl_virtual_meters "
-                 " ORDER BY id ")
-        cursor.execute(query)
+                 " FROM tbl_virtual_meters " )
+        params = []
+        if search_query:
+            query += " WHERE name LIKE %s OR description LIKE %s "
+            params = [f'%{search_query}%', f'%{search_query}%']
+
+        query += " ORDER BY id "
+        cursor.execute(query,params)
         rows_virtual_meters = cursor.fetchall()
 
         result = list()